OCSADV-255 Move Target conversion methods to ModelConverters #time 20m
