+++ bundle/edu.gemini.ags/src/main/scala/edu/gemini/ags/impl/ScienceTargetStrategy.scala
@@ -1,7 +1,7 @@
 package edu.gemini.ags.impl
 
 import edu.gemini.ags.api.AgsMagnitude._
-import edu.gemini.ags.impl._
+import edu.gemini.pot.ModelConverters._
 import edu.gemini.ags.api.{AgsAnalysis, AgsStrategy}
 import edu.gemini.spModel.ags.AgsStrategyKey
 import edu.gemini.spModel.core.MagnitudeBand
