+++ bundle/edu.gemini.pot/src/main/scala/edu/gemini/pot/ModelConverters.scala
@@ -7,6 +7,10 @@ import edu.gemini.spModel.core._
 import edu.gemini.spModel.core.AngleSyntax._
 import edu.gemini.spModel.rich.shared.immutable._
 import edu.gemini.spModel.target.SPTarget
+import edu.gemini.spModel.target.system.HmsDegTarget
+
+import scalaz._
+import Scalaz._
 
 /**
  * This code is a conglomeration of some methods from the edu.gemini.ags package's impl/package.scala file
@@ -130,7 +134,9 @@ object ModelConverters {
     def toOldModel: skyobject.SkyObject = {
       val ra          = skycalc.Angle.degrees(st.coordinates.ra.toAngle.toDegrees)
       val dec         = skycalc.Angle.degrees(st.coordinates.dec.toAngle.toDegrees)
-      val coordinates = new skyobject.coords.HmsDegCoordinates.Builder(ra, dec).build()
+      val coordinates = st.properMotion.map { pm =>
+            new skyobject.coords.HmsDegCoordinates.Builder(ra, dec).pmRa(skycalc.Angle.milliarcsecs(pm.deltaRA.velocity.masPerYear)).pmDec(skycalc.Angle.milliarcsecs(pm.deltaDec.velocity.masPerYear)).build()
+        } |  new skyobject.coords.HmsDegCoordinates.Builder(ra, dec).build()
       val mags        = st.magnitudes.map(_.toOldModel)
       new skyobject.SkyObject.Builder(st.name, coordinates).magnitudes(mags: _*).build()
     }
@@ -144,7 +150,10 @@ object ModelConverters {
       val dec         = Angle.fromDegrees(so.getHmsDegCoordinates.getDec.toDegrees.getMagnitude)
       val coordinates = Coordinates(RightAscension.fromAngle(ra), Declination.fromAngle(dec).getOrElse(Declination.zero))
       val mags        = so.getMagnitudes.asScala.map(_.toNewModel)
-      SiderealTarget(so.getName, coordinates, None, mags.toList, None)
+      val pmRa        = RightAscensionAngularVelocity(AngularVelocity(so.getHmsDegCoordinates.getPmRa.toMilliarcsecs.getMagnitude))
+      val pmDec       = DeclinationAngularVelocity(AngularVelocity(so.getHmsDegCoordinates.getPmDec.toMilliarcsecs.getMagnitude))
+      val pm          = ProperMotion(pmRa, pmDec)
+      SiderealTarget(so.getName, coordinates, Some(pm), mags.toList, None)
     }
   }
 
@@ -156,7 +165,13 @@ object ModelConverters {
       val ra          = Angle.fromDegrees(coords.getRaDeg)
       val dec         = Angle.fromDegrees(coords.getDecDeg)
       val coordinates = Coordinates(RightAscension.fromAngle(ra), Declination.fromAngle(dec).getOrElse(Declination.zero))
-      SiderealTarget(name, coordinates, None, mags, None)
+
+            // Only HmsDegTargets have a proper motion and the values are in milli arcsecs/year
+      val pm          = sp.getTarget match {
+        case t:HmsDegTarget => Some(ProperMotion(RightAscensionAngularVelocity(AngularVelocity(t.getPropMotionRA)), DeclinationAngularVelocity(AngularVelocity(t.getPropMotionDec))))
+        case _              => None
+      }
+      SiderealTarget(name, coordinates, pm, mags, None)
     }
   }
 }
